requirements.txt
strands-agents>=0.1.0
streamlit>=1.24.0
boto3>=1.26.0
Pillow>=9.5.0
moviepy>=1.0.3
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
opencv-python>=4.8.0
requests>=2.31.0